refactoring job package
