+++ src/lib/modules/job/monitor.go
@@ -1,0 +1,90 @@
+package job
+
+import (
+	"bytes"
+	"encoding/json"
+	"fmt"
+	"net/http"
+	"strconv"
+	"strings"
+	"time"
+
+	config "github.com/bench-routes/bench-routes/src/lib/config_v2"
+	"github.com/bench-routes/bench-routes/src/lib/modules/evaluate"
+	"github.com/bench-routes/bench-routes/tsdb/file"
+)
+
+type monitoringJob struct {
+	JobInfo
+	app     file.Appendable
+	sigCh   chan struct{}
+	client  *http.Client
+	request *http.Request
+}
+
+// newMonitoringJob creates a new monitoringJob.
+func newMonitoringJob(app file.Appendable, c chan struct{}, api *config.API) (*monitoringJob, error) {
+	newJob := &monitoringJob{
+		app:   app,
+		sigCh: c,
+		JobInfo: JobInfo{
+			Name:        api.Name,
+			Every:       api.Every,
+			lastExecute: time.Now().Add(api.Every * -1),
+		},
+		client: &http.Client{
+			Timeout: time.Second * 5,
+		},
+		request: &http.Request{},
+	}
+	if err := newReq(newJob, api); err != nil {
+		return nil, fmt.Errorf("error making http request : %w", err)
+	}
+	return newJob, nil
+}
+
+// Execute executes the monitoringJob.
+func (mn *monitoringJob) Execute() {
+	for range mn.sigCh {
+		mn.JobInfo.writeTime()
+		response, err := evaluate.Monitor(mn.client, mn.request)
+		if err != nil {
+			fmt.Println(fmt.Errorf("%s", err))
+		}
+		val := fmt.Sprintf("%v|%v", response.Delay.Microseconds(), strconv.Itoa(response.Length))
+		fmt.Println(val)
+		mn.app.Append(file.NewBlock("monitoring", val))
+	}
+}
+
+// newReq creates http client and request for the job.
+func newReq(job *monitoringJob, api *config.API) error {
+	var body []byte
+	var err error
+	if len(api.Body) != 0 {
+		body, err = json.Marshal(api.Body)
+		if err != nil {
+			return fmt.Errorf("error marshalling : %w", err)
+		}
+	}
+	// creating http request for job.
+	job.request, err = http.NewRequest(strings.ToUpper(api.Method), api.Domain+api.Route, bytes.NewBuffer(body))
+	if err != nil {
+		return fmt.Errorf("error making http request : %w", err)
+	}
+	// adding headers to the request.
+	for k, v := range api.Headers {
+		job.request.Header.Set(k, v)
+	}
+	return nil
+}
+
+// Aborts aborts the running job.
+func (mn *monitoringJob) Abort() {
+	close(mn.sigCh)
+}
+
+// Info returns jobInfo of the job.
+func (mn *monitoringJob) Info() *JobInfo {
+	return &mn.JobInfo
+}
